+++ models/user.js
@@ -10,6 +10,8 @@ module.exports = (sequelize, DataTypes) => {
      * The `models/index` file will call this method automatically.
      */
     static associate(models) {
+      // User.hasMany(FoodDonation, {as: 'FoodDonations'});
+
       // define association here
     }
   };
